requirements.txt
httpx==0.28.1
pywin32
python-dotenv